# docs/swagger.yaml
basePath: /api
definitions:
...
  models.ChannelReq:
    properties:
      metadata:
        $ref: '#/definitions/models.Metadata'
      name:
        example: channel1
        type: string
    required:
    - name
    type: object
  models.ChannelRes:
    properties:
      id:
        example: 880d7429-8857-4e50-a7e0-698e2865b0aa
        type: string
      metadata:
        $ref: '#/definitions/models.Metadata'
      name:
        example: channel1
        type: string
    type: object
  models.ChannelResAll:
    properties:
      id:
        example: 880d7429-8857-4e50-a7e0-698e2865b0aa
        type: string
      metadata:
        $ref: '#/definitions/models.Metadata'
      name:
        example: channel1
        type: string
      owner:
        example: user@example.com
        type: string
    type: object
  models.ChannelsList:
    properties:
      channels:
        items:
          $ref: '#/definitions/models.ChannelResAll'
        type: array
    type: object
...
  models.Metadata:
    additionalProperties: true
    type: object
...
  models.ThingReq:
    properties:
      coordinates:
        additionalProperties: true
        type: object
      metadata:
        $ref: '#/definitions/models.Metadata'
      name:
        example: device1
        type: string
    required:
    - name
    type: object
  models.ThingRes:
    properties:
      id:
        example: 8c0c7129-8857-4e50-a7e0-698e2865b0aa
        type: string
      key:
        example: ef751d71-fb43-423c-a2eb-8602e6232cb4
        type: string
      metadata:
        $ref: '#/definitions/models.Metadata'
      name:
        example: device1
        type: string
    type: object
  models.ThingResAll:
    properties:
      groups:
        items:
          type: string
        type: array
      id:
        example: 8c0c7129-8857-4e50-a7e0-698e2865b0aa
        type: string
      key:
        example: ef751d71-fb43-423c-a2eb-8602e6232cb4
        type: string
      metadata:
        $ref: '#/definitions/models.Metadata'
      name:
        example: device1
        type: string
      owner:
        example: user@example.com
        type: string
    type: object
  models.ThingsPageRes:
    properties:
      dir:
        type: string
      isadmin:
        type: boolean
      limit:
        type: integer
      offset:
        type: integer
      order:
        type: string
      things:
        items:
          $ref: '#/definitions/models.ThingResAll'
        type: array
      total:
        type: integer
    type: object
...
host: iot.discretal.com
info:
  contact: {}
  description: A wrapper api for utilizing Discretal server messaging services over
    MQTT
  termsOfService: http://iot.discretal.com/terms/
  title: Discretal API
  version: "1.0"
paths:
  /channels:
    get:
      description: Retrieves a list of channels. Due to performance concerns, data
        is retrieved in subsets. The API things must ensure that the entire dataset
        is consumed either by making subsequent requests, or by increasing the subset
        size of the initial request.
      parameters:
      - description: Size of the subset to retrieve.
        in: query
        name: limit
        type: integer
      - description: Number of items to skip during retrieval.
        in: query
        name: offset
        type: integer
      - description: Unique channel name.
        in: query
        name: name
        type: string
      - description: Entity to be sorted on.
        in: query
        name: order
        type: string
      - description: Asc or Desc sorting.
        in: query
        name: dir
        type: string
      - description: Disconnected true or false.
        in: query
        name: disconnected
        type: boolean
      - description: Email ID of selected user.
        in: query
        name: email
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Data retrieved.
          schema:
            $ref: '#/definitions/models.ChannelsList'
        "400":
          description: Failed due to malformed query parameters.
        "401":
          description: Missing or invalid access token provided.
        "500":
          description: Unexpected server-side error occurred.
      security:
      - BearerAuth: []
      summary: Retrieves channels
      tags:
      - channels
    post:
      description: Creates new channel. User identified by the provided access token
        will be the channels owner.
      parameters:
      - description: JSON-formatted document describing the updated channel.
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/models.ChannelReq'
      produces:
      - application/json
      responses:
        "201":
          description: Channel created.
          schema:
            $ref: '#/definitions/models.ChannelRes'
        "400":
          description: Failed due to malformed JSON.
        "401":
          description: Missing or invalid access token provided.
        "500":
          description: Unexpected server-side error occurred.
      security:
      - BearerAuth: []
      summary: Adds new channel
      tags:
      - channels
...
  /things:
    get:
      description: Retrieves a list of things. Due to performance concerns, data is
        retrieved in subsets. The API things must ensure that the entire dataset is
        consumed either by making subsequent requests, or by increasing the subset
        size of the initial request.
      parameters:
      - description: Size of the subset to retrieve.
        in: query
        name: limit
        type: integer
      - description: Number of items to skip during retrieval.
        in: query
        name: offset
        type: integer
      - description: Unique thing name.
        in: query
        name: name
        type: string
      - description: Entity to be sorted on.
        in: query
        name: order
        type: string
      - description: Asc or Desc sorting.
        in: query
        name: dir
        type: string
      - description: Disconnected true or false.
        in: query
        name: disconnected
        type: boolean
      - description: Email ID of selected user.
        in: query
        name: email
        type: string
      - description: Array of group IDs.
        in: query
        name: gids
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: Data retrieved.
          schema:
            $ref: '#/definitions/models.ThingsPageRes'
        "400":
          description: Failed due to malformed query parameters.
        "401":
          description: Missing or invalid access token provided.
        "500":
          description: Unexpected server-side error occurred.
      security:
      - BearerAuth: []
      summary: Retrieves things
      tags:
      - things
    post:
      description: Adds new thing to the list of things owned by user identified using
        the provided access token.
      parameters:
      - description: JSON-formatted document describing the new thing.
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/models.ThingReq'
      produces:
      - application/json
      responses:
        "201":
          description: Thing registered.
          schema:
            $ref: '#/definitions/models.ThingRes'
        "400":
          description: Failed due to malformed JSON.
        "401":
          description: Missing or invalid access token provided.
        "500":
          description: Unexpected server-side error occurred.
      security:
      - BearerAuth: []
      summary: Adds new thing
      tags:
      - things
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"